[INLONG-7319][Dashboard] Support global configuration of Provider and Layout
