+++ inlong-dashboard/src/components/DashboardCard/index.module.less
@@ -18,8 +18,6 @@
 * under the License.
 */
 
-@import '@/themes/antd.var.less';
-
 .dashCardNums {
   font-family: @number-font-family;
   font-size: 36px;
