+++ inlong-dashboard/src/index.tsx
@@ -19,6 +19,6 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from '@/router';
+import App from '@/App';
 
 ReactDOM.render(<App />, document.getElementById('root'));
