--- inlong-dashboard/src/themes/style.less
@@ -1,64 +1,0 @@
-
-/*
-* Licensed to the Apache Software Foundation (ASF) under one
-* or more contributor license agreements.  See the NOTICE file
-* distributed with this work for additional information
-* regarding copyright ownership.  The ASF licenses this file
-* to you under the Apache License, Version 2.0 (the
-* "License"); you may not use this file except in compliance
-* with the License.  You may obtain a copy of the License at
-*
-*   http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6 {
-  margin: 0;
-}
-
-body {
-  // https://github.com/postcss/postcss-color-function
-  background: #fff;
-  @media (--abs-md) {
-    background-color: #333;
-  }
-}
-
-::-webkit-scrollbar {
-  width: 8px;
-  height: 8px;
-  background: none;
-}
-
-::-webkit-scrollbar-thumb {
-  border-radius: 4px;
-  background-color: #ccc;
-}
-
-::-webkit-scrollbar-corner {
-  background-color: #eaedf0;
-}
-
-ul,
-ol {
-  list-style: none;
-}
-
-pre {
-  padding: 10px;
-  background: #f6f7fb;
-  font-size: 12px;
-  overflow: auto;
-  border-radius: 8px;
-}
